requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.1.7
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cycler==0.10.0
Django==3.0.3
django-allauth==0.41.0
django-audiotracks==0.2.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
hashids==1.2.0
idna==2.9
importlib-metadata==1.7.0
joblib==0.15.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
mutagenx==1.22
oauthlib==3.1.0
numpy==1.18.4
opt-einsum==3.2.1
pandas==1.0.3
phonenumbers==8.12.6
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
termcolor==1.1.0
threadpoolctl==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
